wip: consolidate markdown template logic
